Fix bug with negative traces
Casting the mask to uint8 dtype when binarizing a mask appears to
result in weird dypte related under or overflows when extracting the
trace. This results in negative numbers in the trace which should
never happen for this kind of data modality. Removing the dtype
cast that occurs in the mask binarization function appears to
resolve the issue.