GH #10: Feature Extraction and Inference module
Creates the 'classification' module for feature
extraction and inference. Given the proper input data,
and a model, the module will generate predictions
for whether an ROI is a cell or not. If the ROI is
not a cell, it will be marked invalid and an exclusion
label will be added to the ROI's 'exclusion_labels'
field. Note that the current exclusion label is a
placeholder (see GH #16).

The module uses argschema, so can be run with
an input json or command line arguments.

Currently the neuropil traces are unused because
croissant's FeatureExtractor does not yet support them,
but the infrastructure is already there once they are
available.