Gh8/feature/cell binarization and creation module (#14)
* Added ROI Mask Binarization and Cell Creation

Adds a module, convert_rois, that takes suite2p source extraction outputs in the form of stat.npy files and binarizes the masks contained within. The masks are converted to LIMs compatible format and motion exclusion is calculated for each ROI given motion border data provided to this module as input. The ROIs are written out in .json format to the specified location.