Fix classifier module inability to load AWS S3 URIs
Although the classifier module currently checks that
an S3 URI is valid with its input schema, it doesn't
actually 'download' the S3 file to a local path which
joblib requires in order to `.load()`.

This commit adds a `load_model()` function so that
these S3 URIs can be downloaded to a temporary file and
loaded by joblib.

This commit also replaces the `valid_model_registry_env`
input schema method with the marshmallow OneOf() built-in
validator.