Fix mismatch between roi id and trace id when loading traces
The AllenSDK trace extraction module performs a sort by
ROI mask 'ids' before performing trace extraction:

allensdk.brain_observatory.ophys.trace_extraction.create_roi_masks

As a result, traces are not guaranteed to be in the same order
as the ROIs that were used to create the traces. Luckily,
the trace files (*.h5) that are produced have an additional key called
"roi_names" which does relate each trace with an ROI "id".

This commit changes how traces are loaded in the ophys_etl
classification module so that the order of traces after loading  will
match up with the order of ROIs.