Fix lims and module ROI id mismatch
Previously, the classification _munge_traces() function
was assuming that ROI ids would take on a contiguous range
of values starting at 0. Because ROI id names are now
being obtained from LIMS, it is no longer the case that
ROI ids will start from 0. It is also very likely that
the expectation that ROI ids be contiguous is also no
longer the case.

This commit fixes the mapping that is used to match up
the index which a trace is stored with an ROI id.