rewrite _flag_to_events
It does not look like the original code was doing what was
advertised (selecting all of the active elements in a trace
plus a buffer of len_ne timestamps to either side of each event).
This method is tested against ground truth and is,
I think, easier to maintain.

This change necessitates removing some of the tests that just compare the
output of active_traces.py to output from the prototype code, since
the prototype code was not quite doing what we wanted.

I also needed to add a patch to decrosstalk_utils.get_crosstalk_data
to handle the case where an inactive trace was passed in.