increase verbosity of pytest invocation
In February 2022, some tests were failing because multiprocessing
was not cleaning itself up properly on CircleCI. This problem was
not evident from the log until these flags were added to our
CircleCI configuration script. "Unfortunately" the problem resolved
itself. We are adding these flags here so we can more easily
diagnose the problem (we hope), should it arise again.