run all tests in a Docker container that we create
The reasons for this change are

- Running the suite2p tests in the Docker containers provided by CircleCI
was an order of magnitude slower than running them on a "machine" instance
provided by CircleCI and building our own Docker containers. I cannot say
for certain why this was the case (the "medium" version of CircleCI's
"machine" instance is larger than the "medium" version of CircleCI's
Docker executors, but even going to a "large" Docker executor did
not improve performance enough)

https://circleci.com/docs/2.0/configuration-reference/#resourceclass

- pytest-cov did not interact well with multiprocessig. There were
frequently failures in which pytest-cov could not write to the sql database
used to track code coverage. Curiously, these failures did not cause test
failures. They just occurred silently, resulting in incorrect code coverage
reports. See ~ line 156 of

https://app.circleci.com/pipelines/github/AllenInstitute/ophys_etl_pipelines/2347/workflows/fad64973-bbad-4304-8dd7-ac3c91e25228/jobs/7634

that is why this commit does away with the pytest-cov plug in and uses
coverage.py directly