remove unused invocations of np.max
The second arg of np.max is actually an axis, so these
calls to np.max were not clipping the max shifts to be >= 0
(in fact, if you look at tests/utils/test_motion_border.py,
our unit tests were written to expect negative values as a
legal output).

Probably this was not intended, but, all of our code has
been built around the current implementation, so I'm not going
to alter the logic in get_max_correction_values now.

Ticket #458 has been opened to do a more thorough and
self-consistent investigation of our treatment of motion borders
in the future. At that time, we can consider altering
get_max_correction_values to only return values >= 0